+++ iOS/Projects/App/WeTri/Sources/CommonScene/Coordinator/AppCoordinator.swift
@@ -8,6 +8,7 @@
 
 import Coordinator
 import LoginFeature
+import SignUpFeature
 import SplashFeature
 import UIKit
 
@@ -49,8 +50,22 @@ final class AppCoordinator: AppCoordinating {
   }
 
   func showLoginFlow() {
-    // TODO: LoginCoordinator 연결
-    let coordinator = LoginFeatureCoordinator(navigationController: navigationController)
+    let coordinator = LoginFeatureCoordinator(
+      navigationController: navigationController,
+      isMockEnvironment: false,
+      isMockFirst: true
+    )
+    childCoordinators.append(coordinator)
+    coordinator.finishDelegate = self
+    coordinator.start()
+  }
+
+  func showSignUpFlow(newUserInformation: NewUserInformation) {
+    let coordinator = SignUpFeatureCoordinator(
+      navigationController: navigationController,
+      newUserInformation: newUserInformation,
+      isMockEnvironment: false
+    )
     childCoordinators.append(coordinator)
     coordinator.finishDelegate = self
     coordinator.start()
@@ -71,6 +86,7 @@ extension AppCoordinator: CoordinatorFinishDelegate {
     childCoordinators = childCoordinators.filter {
       $0.flow != childCoordinator.flow
     }
+    navigationController.popToRootViewController(animated: false)
   }
 }
 
@@ -82,6 +98,33 @@ extension AppCoordinator: SplashCoordinatorFinishDelegate {
       showLoginFlow()
     } else {
       showTabBarFlow()
+    }
+  }
+}
+
+// MARK: SignUpFeatureCoordinatorFinishDelegate
+
+extension AppCoordinator: SignUpFeatureCoordinatorFinishDelegate {
+  func signUpFeatureCooridnatorDidFinished() {
+    showTabBarFlow()
+  }
+}
+
+// MARK: LoginFeatureFinishDelegate
+
+extension AppCoordinator: LoginFeatureFinishDelegate {
+  func loginFeatureCoordinatorDidFinished(initialUser: InitialUser?, token: Token?) {
+    if let initialUser {
+      // TODO: 처음접속하는 유저일 경우 signUp
+      showSignUpFlow(newUserInformation: NewUserInformation(
+        mappedUserID: initialUser.mappedUserID,
+        provider: initialUser.provider
+      )
+      )
+    }
+
+    if let token {
+      // TODO: 기존에 접속해본적이 있어서 로그인하고 토큰을 받은 경우
     }
   }
 }
