+++ iOS/Projects/Features/Login/Sources/Domain/Entity/InitialUser.swift
@@ -11,15 +11,25 @@ import Foundation
 // MARK: - InitialUser
 
 /// 처음 로그인 하는 유저의 Response를 담을 Entity
-struct InitialUser {
+public struct InitialUser {
   /// 처음 로그인 하는지 아닌지
-  let isFirstLogined: Bool
+  public let isFirstLogined: Bool
 
   ///
-  let mappedUserID: String
+  public let mappedUserID: String
 
   /// OAuth 로그인 종류
-  let provider: AuthProvider
+  public let provider: AuthProvider
+
+  public init(
+    isFirstLogined: Bool,
+    mappedUserID: String,
+    provider: AuthProvider
+  ) {
+    self.isFirstLogined = isFirstLogined
+    self.mappedUserID = mappedUserID
+    self.provider = provider
+  }
 }
 
 // MARK: Codable
@@ -28,6 +38,6 @@ extension InitialUser: Codable {}
 
 // MARK: - AuthProvider
 
-enum AuthProvider: String, Codable {
+public enum AuthProvider: String, Codable {
   case apple
 }
