+++ iOS/Projects/Features/Login/Sources/Domain/Entity/Token.swift
@@ -11,14 +11,14 @@ import Foundation
 // MARK: - Token
 
 /// 백엔드로부터 받아온 JWT를 담을 데이터 entity
-struct Token {
+public struct Token {
   /// accessToken
   let accessToken: String?
 
   /// refreshToken
   let refreshToken: String?
 
-  init(accesToken: String? = nil, refreshToken: String? = nil) {
+  public init(accesToken: String? = nil, refreshToken: String? = nil) {
     accessToken = accesToken
     self.refreshToken = refreshToken
   }
