--- iOS/Projects/Features/Login/Sources/Presentation/LoginScene/Coordinator/Delegate/LoginDidFinishedDelegate.swift
@@ -1,13 +1,0 @@
-//
-//  LoginDidFinishedDelegate.swift
-//  LoginFeature
-//
-//  Created by 안종표 on 12/7/23.
-//  Copyright © 2023 kr.codesquad.boostcamp8. All rights reserved.
-//
-
-import Foundation
-
-protocol LoginDidFinishedDelegate: AnyObject {
-  func loginCoordinatorDidFinished(initialUser: InitialUser?, token: Token?)
-}
