+++ iOS/Projects/Features/Login/Sources/Presentation/LoginScene/Coordinator/Protocol/LoginFeatureCoordinating.swift
@@ -11,4 +11,5 @@ import Foundation
 
 protocol LoginFeatureCoordinating: Coordinating {
   func showLoginFlow()
+  func finish(initialUser: InitialUser?, token: Token?)
 }
