+++ iOS/Projects/Features/SignUp/Sources/Presentation/Common/Coordinator/Protocol/SignUpFeatureCoordinating.swift
@@ -10,5 +10,5 @@ import Coordinator
 import Foundation
 
 public protocol SignUpFeatureCoordinating: Coordinating {
-  func showSignUpFlow()
+  func pushSingUpContainerViewController()
 }
